+++ lib/Data/Time/Calendar/CalendarDiffDays.hs
@@ -12,6 +12,7 @@ import Data.Semigroup hiding (option)
 #endif
 import Data.Typeable
 import Data.Data
+import Control.DeepSeq
 
 data CalendarDiffDays = CalendarDiffDays
     { cdMonths :: Integer
@@ -26,6 +27,9 @@ data CalendarDiffDays = CalendarDiffDays
     -- ^ @since 1.9.2
 #endif
     )
+
+instance NFData CalendarDiffDays where
+    rnf (CalendarDiffDays m d) = rnf m `seq` rnf d `seq` ()
 
 -- | Additive
 instance Semigroup CalendarDiffDays where