+++ lib/Data/Time/Calendar/Week.hs
@@ -10,7 +10,9 @@ module Data.Time.Calendar.Week
     ) where
 
 import Data.Fixed
+import Data.Ix
 import Data.Data
+import Control.DeepSeq
 import Data.Time.Calendar.Days
 
 data DayOfWeek
@@ -21,7 +23,16 @@ data DayOfWeek
     | Friday
     | Saturday
     | Sunday
-    deriving (Eq, Show, Read, Data, Typeable, Ord)
+    deriving (Eq, Show, Read, Data, Typeable, Ord, Ix)
+
+instance NFData DayOfWeek where
+    rnf Monday = ()
+    rnf Tuesday = ()
+    rnf Wednesday = ()
+    rnf Thursday = ()
+    rnf Friday = ()
+    rnf Saturday = ()
+    rnf Sunday = ()
 
 -- | \"Circular\", so for example @[Tuesday ..]@ gives an endless sequence.
 -- Also: 'fromEnum' gives [1 .. 7] for [Monday .. Sunday], and 'toEnum' performs mod 7 to give a cycle of days.