+++ changelog.md
@@ -3,7 +3,7 @@
 ## [1.11.1]
 - all modules Safe or Trustworthy
 - fix NFData instances for DiffTime, NominalDiffTime, TimeOfDay
-- add Ix, Enum and NFData instance to Month, Quarter and QuarterOfYear types
+- add missing Ix, Enum, NFData instances to DayOfWeek, CalendarDiffDays, CalendarDiffTime, Month, Quarter, QuarterOfYear
 
 ## [1.11]
 - new calendrical type synonyms and abstract constructors